+++ lib/buildcontext/navigation.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/widgets.dart';
+
+extension BuildContextNavigationEx on BuildContext {
+  NavigatorState get navigator => Navigator.of(this);
+}