+++ lib/widget/container.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/cupertino.dart';
+
+extension ContainerWidgetEx on Widget {
+  Container container({
+    Key? key,
+    AlignmentGeometry? align,
+    EdgeInsetsGeometry? pad,
+    Color? color,
+    Decoration? decoration,
+    Decoration? foregroundDecoration,
+    double? width,
+    double? height,
+    BoxConstraints? constraints,
+    EdgeInsetsGeometry? margin,
+    Matrix4? transform,
+    AlignmentGeometry? transformAlign,
+    Clip clipBehavior = Clip.none,
+  }) =>
+      Container(
+        key: key,
+        alignment: align,
+        padding: pad,
+        color: color,
+        decoration: decoration,
+        foregroundDecoration: foregroundDecoration,
+        width: width,
+        height: height,
+        constraints: constraints,
+        margin: margin,
+        transform: transform,
+        transformAlignment: transformAlign,
+        clipBehavior: clipBehavior,
+        child: this,
+      );
+}